9.4. Рефакторинг. Разбивает app.exception-filter
