+++ src/shared/libs/rest/exception-filter/app.exception-filter.ts
@@ -1,0 +1,26 @@
+import { inject, injectable } from 'inversify';
+import { StatusCodes } from 'http-status-codes';
+import { NextFunction, Request, Response } from 'express';
+
+import { createErrorObject } from '../../../helpers/index.js';
+
+import { ExceptionFilter } from './exception-filter.interface.js';
+import { Logger } from '../../logger/index.js';
+import { COMPONENT_MAP } from '../../../types/index.js';
+import { ApplicationError } from '../index.js';
+
+@injectable()
+export class AppExceptionFilter implements ExceptionFilter {
+  constructor(
+    @inject(COMPONENT_MAP.LOGGER) private readonly logger: Logger
+  ) {
+    this.logger.info('Register AppExceptionFilter');
+  }
+
+  public catch(error: Error, _req: Request, res: Response, _next: NextFunction): void {
+    this.logger.error(error.message, error);
+    res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json(createErrorObject(ApplicationError.ServiceError, error.message));
+  }
+}
