+++ src/shared/libs/rest/exception-filter/validation.exception-filter.ts
@@ -3,16 +3,16 @@ import { NextFunction, Request, Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
 
 import { ExceptionFilter } from './exception-filter.interface.js';
-import { Component } from '../../../types/index.js';
 import { Logger } from '../../logger/index.js';
 import { ValidationError } from '../errors/index.js';
 import { createErrorObject } from '../../../helpers/index.js';
 import { ApplicationError } from '../types/application-error.enum.js';
+import { COMPONENT_MAP } from '../../../types/component-map.enum.js';
 
 @injectable()
 export class ValidationExceptionFilter implements ExceptionFilter {
   constructor(
-    @inject(Component.Logger) private readonly logger: Logger
+    @inject(COMPONENT_MAP.LOGGER) private readonly logger: Logger
   ) {
     this.logger.info('Register ValidationExceptionFilter');
   }
