+++ src/shared/libs/rest/index.ts
@@ -2,6 +2,8 @@ export { RequestParams } from './types/request-params.type.js';
 export { RequestBody } from './types/request-body.type.js';
 export { HttpMethod } from './types/http-method.enum.js';
 export { ValidationErrorField } from './types/validation-error-field.type.js';
+export { ValidationExceptionFilter } from './exception-filter/validation.exception-filter.js';
+export { HttpErrorExceptionFilter } from './exception-filter/http-error.exception-filter.js';
 export { ApplicationError } from './types/application-error.enum.js';
 export { Route } from './types/route.interface.js';
 export { Middleware } from './middleware/middleware.interface.js';
@@ -19,6 +21,6 @@ export { AuthController } from './controller/auth.controller.js';
 
 
 export { ExceptionFilter } from './exception-filter/exception-filter.interface.js';
-export { AppExceptionFilter } from './exception-filter/app-exception-filter.js';
+export { AppExceptionFilter } from './exception-filter/app.exception-filter.js';
 
 export { HttpError } from './errors/index.js';
