+++ src/shared/types/component-map.enum.ts
@@ -16,4 +16,6 @@ export const COMPONENT_MAP = {
   AUTH_CONTROLLER: Symbol(),
   EXCEPTION_FILTER: Symbol(),
   AUTH_EXCEPTION_FILTER: Symbol(),
+  HTTP_EXCEPTION_FILTER: Symbol(),
+  VALIDATION_EXCEPTION_FILTER: Symbol(),
 } as const;
